requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
numpy==1.20.1
pandas==1.2.3
paretoset==1.2.0
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
wincertstore==0.2
